update phan quyen front-end

+++ WEBBANSACH-BACKEND/WEBBANSACH-BACKEND/src/main/java/com/vn/webbansach_backend/entity/HinhThucThanhToan.java
@@ -20,7 +20,7 @@ public class HinhThucThanhToan {
     private int maHinhThucThanhToan;
     private String tenHinhThucThanhToan;
     private String moTa;
-    private double chiPhiThanhToan;
+    private Long chiPhiThanhToan;
 
     @OneToMany(mappedBy = "hinhThucThanhToan", cascade = {CascadeType.PERSIST, CascadeType.DETACH, CascadeType.MERGE, CascadeType.REFRESH})
     private List<DonHang> danhSachDonHang;
